Move interval queries out into a concern
